+++ src/click_repl/__init__.py
@@ -5,10 +5,10 @@ inside your click app in just some simple steps.
 
 from __future__ import annotations
 
-from ._internal_command import repl_exit
 from ._repl import Repl, ReplCli, repl
 from .core import ReplContext
 from .decorators import pass_context, register_repl
+from .internal_commands import repl_exit
 
 __all__ = [
     "repl_exit",