+++ tests/test_internal_cmds/test_exec_internal_cmds.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 import click
 import pytest
 
-from click_repl._internal_command import InternalCommandSystem
 from click_repl.core import ReplContext
 from click_repl.exceptions import ExitReplException, PrefixNotFound
+from click_repl.internal_commands import InternalCommandSystem
 
 
 @click.command()