+++ tests/test_internal_cmds/test_get_internal_cmds.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import pytest
 
 import click_repl
-from click_repl._internal_command import InternalCommandSystem
-from click_repl._internal_command import repl_exit as repl_exit
+from click_repl.internal_commands import InternalCommandSystem
+from click_repl.internal_commands import repl_exit as repl_exit
 
 internal_command_system = InternalCommandSystem()
 
@@ -21,4 +21,4 @@ def test_get_registered_target_with_default_value(test_input):
 
 def test_exit_repl_function():
     with pytest.raises(click_repl.exceptions.ExitReplException):
-        click_repl._internal_command.repl_exit()
+        click_repl.internal_commands.repl_exit()