+++ tests/test_repl_ctx.py
@@ -4,8 +4,8 @@ import click
 import pytest
 
 import click_repl
-from click_repl._internal_command import InternalCommandSystem
 from click_repl.core import ReplContext
+from click_repl.internal_commands import InternalCommandSystem
 from tests import mock_stdin