+++ docs/source/changelog.rst
@@ -1,7 +1,6 @@
 Change History
 ==============
 
-.. _v0.3.0:
 
 v0.3.0
 ------
@@ -10,7 +9,7 @@ v0.3.0
 | **Release-by**: `Asif Saif Uddin <https://github.com/auvipy>`_
 
 - Drop Python 2 support, removed :mod:`~six`. :issue:`78`
-- Uses :class:`~prompt_toolkit.PromptSession` class from prompt_toolkit instead of :func:`~prompt_toolkit.prompt`` function (:issue:`63`).
+- Uses :class:`~prompt_toolkit.PromptSession` class from prompt_toolkit instead of :func:`~prompt_toolkit.prompt` function (:issue:`63`).
 - Added filter for hidden commands and options (:issue:`86`).
 - Added click's autocompletion support (:issue:`88`).
 - Added tab-completion for :class:`~click.Path` and :obj:`~click.BOOL` type arguments (:issue:`95`).
@@ -19,7 +18,6 @@ v0.3.0
 - Updated completer class and tests based on new fix :issue:`92` (:issue:`102`).
 - Python 3.11 support.
 
-.. _v0.2.0:
 
 v0.2.0
 ------